[SW-0] accel/habanalabs: return ENOMEM if less than requested pages were pinned [2]
EFAULT is currently returned if less than requested user pages are
pinned.
This value means a "bad address" which might be confusing to user, as
the address of the given user memory is not necessarily "bad".

Modify the return value in to ENOMEM, as "out of memory" is more
suitable in this case.

In addition, EFAULT is also returned if frame_vector_to_pages() fails,
instead of using its return value, so modify it as well.

*** SHOULDN'T BE UPSTREAMED ***
(this code path is for older kernel versions)

Change-Id: I199b211fb6f90a2c8fb6498afb3df4b18cdaaabb
Signed-off-by: Tomer Tayar <ttayar@habana.ai>