[FSW-4677] accel/habanalabs/cn: update hl_cn_get_port_state with new struct
As part of the auxiliary update of get_port_state, which fills a struct
with more data on the link state, update the ioctl struct as well.

As a temporary step, call the get_port_state_temp aux func. Later on this
will be replaced with an updated get_port_state aux func.

Change-Id: I18a238e301557e2b1449a99b290cfe57673c3684
Signed-off-by: David Meriin <dmeriin@habana.ai>