[SW-187526] accel/habanalabs/gaudi3: use critical event indication from FW
In certain cases, FW initiates a hard reset without waiting for a
request from the driver.
These cases should be known in the driver side, and maintaining these
conditions in both sides can sometimes lead to discrepancies.
E.g., the driver will assume that a reset is done by FW while it is not,
or the driver won't be aware of such a reset.

To prevent such issues and to simplify things, a critical event
indication is added in the EQ header, which will be used by FW to
indicate that it is going to initiate a reset.

Remove the hard-coded places in the driver that handle this reset, and
replace it with a single common check of the new indication.

Change-Id: I84a381c858cc40baf3fa5d1a7b341868a0ed77b6
Signed-off-by: Tomer Tayar <ttayar@habana.ai>