[SW-187006] accel/habanalabs/gaudi2: read preboot status after recovering from dirty state
Dirty state can happen when the host VM undergoes a reset, and the
device not.
In such case, the driver has to reset the device before it can be
used again.
As it undergoes reset, device capabilities are zeroed hence we
must read the Preboot's status again to learn Preboot's state,
capabilities, and security.

Change-Id: I865e1a73a08188d73d650ab3bcac73401246859b
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>