[SW-190657] accel/habanalabs/gaudi3: modify handling of EQ data size mismatch
Currently a mismatch between the expected EQ data size and the actual
size is treated as an event without a valid cause.
While this was good for the integration effort with the FW, now that
things are stable, the driver should expect a possible size mismatch, as
new fields can be added, and driver can work with old and new FW
versions.

Modify the handling to accept such mismatches, and move the
responsibility of ignoring new fields, when working with old FW, to the
specific EQ handler.

Change-Id: I682e0dcf00c3975eceabd206ba0a51c8515455c3
Signed-off-by: Tomer Tayar <ttayar@habana.ai>