[SW-187006] accel/habanalabs/gaudi3: read preboot status after recovering from dirty state
Dirty state can happen when the host VM undergoes a reset, and the
device not.
In such case, the driver has to reset the device before it can be
used again.
As it undergoes reset, device capabilities are zeroed hence we
must read the Preboot's status again to learn Preboot's state,
capabilities, and security.

Change-Id: Ib49fef68f0e73412b43c3188dce3f4b0db01a487
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>