[SW-189267] makefile: use exported symbols of OFED modules
ib_umem_* functions of the ib_core module are used when building with
the makefile 'importer' target.
On a setup with OFED installed, the driver is built against symbols of
the installed kernel, but because the OFED modules are loaded,
insmod-ing the driver will fail due to "unknown symbols".

Add a check in makefile for the existence of OFED, and if exists, use
the exported symbols of the OFED modules.

Change-Id: Id49cebf77dad4f45e49243fb963c3bc9e9c00ad2
Signed-off-by: Tomer Tayar <ttayar@habana.ai>