[SW-0] accel/habanalabs/gaudi3: fix CQ mode params for CBC invalidation
As part of setting the CQ mode params for CBC invalidation, the pointer
to the job name string is also set, but it is used outside in the
calling function where this string is no longer valid.

Following other places that set CQ mode params in a dedicated function,
exclude the string pointer initialization from this function, and do it
in the calling function.

Change-Id: I00aa7196c0691c7c5069098d462dab28cae4f7af
Signed-off-by: Tomer Tayar <ttayar@habana.ai>