[SW-0] accel/habanalabs: add comment for using ctx after refcount decrement
In hl_release_dmabuf(), 'ctx' is used after calling hl_ctx_put(), to
point to the compute device file.
While it is guaranteed that this hl_ctx_put() call won't be last,
as the device release is prevented as long as a dma-buf object is alive,
it can be still confusing.
Add a comment for why it is still valid to use 'ctx' at that point.

Change-Id: I23ca06b131675a0ef940644264928e66055ed685
Signed-off-by: Tomer Tayar <ttayar@habana.ai>