[SW-0] accel/habanalabs/gaudi3: move the dup init after fw load
The STLB dup init should be done before the fw config as some configs
expects the dup to be ready.
The issue here is running with security,
as at preboot stage the dup block is protected and cannot be accessed
by the driver.
This patch will just move the dup init call to be after loading the fw,
and right before the fw config, as after loading the fit fw is handling
security settings.

Change-Id: I43972e7caecda8f2a9c83d9344c651fcfe2a5135
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>