[SW-0] accel/habanalabs: failure setting host date/time should not be fatal
There are cases where we don't have compatibility when requesting
hl_fw_set_host_date_and_time().

For example, when flash version (which is checked in hl_fw_version_cmp())
differs from the preboot's.

Failure to get this data should not be fatal but rather be reported as
an info log to the user.

Change-Id: I0272a527c6d184d19310e747da8f9e2fbb9a54f2
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>