[SW-199349] accel/habanalabs: modify CPUCP packets flow
In previous commits, we had a compatibility issue where the SPI flash
version wasn't aligned with the FIT version, hence we got an error during
boot due to an unsupported CPUCP packet.

The W/A at the time was to avoid an error in case of failure in
set_host_data_and_time().

This commit instead is storing the CPUCP version and makes compatibility
decisions based on this version and this way we can avoid any CPUCP
version misalignments.

To achieve this this commit is doing the following:
- remove previous W/A
- use CPUCP version for compatibility decisions in CPUCP packets
- reorder the code such that:
  1. call gaudi3_init_cpu_queues() to set the address of the queues to the
     FW (so we can pass messages).
  2. call cpucp_info_get() this among other things, stores the CPUCP
     version for CPUCP message compatibility
  3. send initial CPUCP messages to the FW.

Change-Id: I68481203abd7445fd48c1decc31a6763a3c373b9
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>