[PNP-190] accel/habanalabs: stall soft reset on status ioctl
Some applications are polling the device status before accessing the
memory using the PCIe link.

Because polling the status and then accessing the device is not atomic,
in some rare cases a soft reset is triggered while accessing the device
after the application polled the device status.

In some case this causes a drain error and the device does a hard reset.

To ensure we can't get a soft reset after polling the status, we suggest
to add an optional stall of 30ms to any soft reset request after polling
the device status.

Change-Id: I62fac579cee0c59ea9c6eb6a56d0db0a2af9de6b
Signed-off-by: Yaniv Levinsky <ylevinsky@habana.ai>