[SW-0] accel/habanalabs/gaudi3: skip checking dbg counter in secured FW
FW is doing the halting protocol upon reset.
The mstr_if block are privileged and driver cannot access
it when running with secured FW.
The debug counters are inside the mstr_if block.
As the FW will make sure no in-flight transactions
during the halt process, so we can safely skip this dbg counter
check in driver.
For non secured fw we can just keep this check for any future need,
won't harm.

Change-Id: I7ae9728af52a4cc024dee27a6f568172626683b8
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>