[SW-196629] accel/habanalabs: add GAUDI3D revision support bugfix4
Add identification to the driver for gaudi3 revision D.

BugFix4:
The bug was that the driver now has limited only for two rev IDs
(1 and 4), on real world include emulations
we could have a lots of revision IDs for each emulation.
To fix this change the driver to set Gaudi3D only when Rev
id 4 is presented, and make the default as Gaudi3 device
instead of invalid.
In addition add missing gaudi3D simulations for the rev id set.

Change-Id: Ia6907f23234bfce18668c1605925a63bd05971c6
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>