[FSW-7088] accel/habanalabs/cn: use new api for hbl_cn_aux_device_info
As part of API change required in FS whereby there should be no asic
specific parameters in aux_data, most of these parameters are moved
to device info instead. This commit sets these parameters in device
info, after previous commits that added the necessary API in the
relevant structs and in other repositories.

Signed-off-by: Sharley Calzolari <sharley.calzolari@intel.com>