[SW-216240] accel/habanalabs/gaudi2: rework override ports masks flow
Upon device registration to the driver there are several code flows which
can require us to modify the ports masks, whether for debug purproses or
when working without FW. In this cases we might need to override the ports
masks, both of the enable ports and the external ports.

This commit addresses this needed and solves a sitation where coral was
loaded with certain card configuration, but as there are no FW and driver
was loaded without module params, the port masks didn't represent the
expected configuration as there were no flow that modifies it.

Signed-off-by: Andrey Agranovich <andrey.agranovich@intel.com>