[FSW-7088] accel/habanalabs/gaudi: implement reserve_irqs api
This new API allows CN to request from compute to reserve a certain
number of IRQs. This will allow compute to manage better the IRQs
that are in use.
In Gaudi1, the implementation of the irq mechanism is different from
newer generations, and since we want to change as little as possible
in Gaudi1, the implementation is empty.

Signed-off-by: Sharley Calzolari <sharley.calzolari@intel.com>