[SW-215834] accel/habanalabs: fix kernel and user ASID allocation (#252)
hl_cn_get_vm_info is returning the same ASID for all vm handles. This
is incorrect in principle as it can be used both when querying user
VM info and driver VM info. The user and driver should have different
ASIDs. However, currently for gaudi the user asid is used throughout.
Therefore, we do not change the current behaviour, but make the code
more aligned with the idea that the asid returned in hl_cn_get_vm_info
depends on the provided vm_handle, such that future extensions will be
easier to implement.

Signed-off-by: Yuval Elmaliah <yuval.elmaliah@intel.com>