[SW-175419] accel/habanalabs/common: modify unwind ladder in hl_init
Current unwind ladder follows the pattern where the labels are named
according to the cleanup action performed. This creates a dependency
issue that makes it hard to pull out "rungs" in the middle: deleting
a certain initialization will require moving labels around.
This situation becomes even worse if some initializations are to be
performed under conditional compilation.

Eliminate this dependency by naming the labels according to the action
that failed. Credit to Lucas De Marchi for the idea.

This patch is done as a cleanup preparation for conditional removal
of non-production code.

Signed-off-by: Ilia Levi <ilia.levi@intel.com>