[SW-224358] accel/habanalabs/common: FW load fail with no reason
Fixed a bug in the driver incorrectly handling ARC memory scrub
failures.
The driver uses an erroneous mask to check for scrub errors and
therefore does not log such errors but correctly checks this error
when checking for FW load failures, which resulted
driver load failures with no apparent reason.

Signed-off-by: Moti Haimovski <mhaimovski@habana.ai>