[SW-227088] accel/habanalabs: implement preboot status verification W/A for device init
W/A for cases where at host reboot PRSTN signal is
not sent to device, hence only the host is being reset.
In such cases, when the host tries to load the driver
it fails on polling for preboot, since it has already
done.

This commit introduces the function that verifies the
preboot status, and should be reverted once fw implements
this solution.
Signed-off-by: Ariel Suller <asuller@habana.ai>