[SW-227088] accel/habanalabs/gaudi3: verify preboot is in valid state at init
W/A for cases where at host reboot PRSTN signal is
not sent to device, hence only the host is being reset.
In such cases, when the host tries to load the driver
it fails on polling for preboot, since it has already
done.

This commit adds a verification at Gaudi3's early init that
preboot boot process is done.
This commit should be reverted once fw implements
this solution.

Signed-off-by: Ariel Suller <asuller@habana.ai>