[SW-0] accel/habanalabs/cn: change stop prepare fw_reset to hard_reset
This flag is passed into the nic driver, and only affects the behavior
of QPC access - if the flag is set - QPC accesses become non-blocking.

In some rare cases if the driver performs a hard-reset while some user
process is using the NIC, it will try to access the QPC and fail.
This flag will prevent such behavior.

In general, it's best if we refactor this mechanism, and remove the
auxiliary devices completely before performing a hard-reset.

Signed-off-by: Tal Risin <tal.risin@intel.com>