[SW-228083] net: hbl_cn: gaudi3: Coalescing of SPI RXB events
SPI RXB events are coalesced in FW and sent to the NIC driver
to avoid SPI interrupt storm. The number of events coalesced
is added by FW in the CPUCP packet.

To ensure FW only enables this coalescing feature for drivers
with version 1.23.0 or greater, driver sends its version details
using CPUCP to FW.
Driver only sends this if the FW itself has a version 1.23.0 or greater
so that it can process this packet properly.

Signed-off-by: Abhijit Chatterjee <abhijit.chatterjee@intel.com>