[SW-220228] accel/habanalabs: receive tx taps from fw and pass to cn
For Rack Scale projects, due to PHY complexity, we need different TX
taps values for each lane depending on cable connected to it.
This information is stored in an EEPROM that the FW has access to.
This commit implements the retrieval of the TX taps from FW and
setting them in cn ports info aux struct for cn's use.

Signed-off-by: Sharley Calzolari <sharley.calzolari@intel.com>