[SW-234764] accel/habanalabs/common: restrict debugfs file permissions
Some debugfs file entries in the habanalabs driver may expose
user-specific or sensitive data. To prevent potential information
leaks, this commit tightens the file permissions of debugfs entries
by limiting access to the file owner (typically root) and group only,
and in some cases to the owner of the file only.
This ensures that only privileged users can read or write these files.

Signed-off-by: Moti Haimovski <motihaim@intel.com>