[VLK-81142] accel/habanalabs: improve EWR enablement logic
Refactored the activation logic for Early-Write-Response (EWR) to
ensure it is only enabled when a DMA buffer is attached to another
device.
Added a use count was added to manage run-time disabling of the
feature, ensuring optimal feature usage.
Enhanced querying of device support of EWR from firmware
Added a lush of all pending writes before enabling or disabling EWR.

Signed-off-by: Moti Haimovski <moti.haimovski@intel.com>